requirements.txt
uvicorn
fastapi
langchain
langchain-community 
langchain-groq
pydantic
typing
python-dotenv
streamlit
